Better organization, added assert_prop
