+++ pecan/lang/pecan_ast/__init__.py
@@ -2,4 +2,5 @@ from pecan.lang.pecan_ast.arith import *
 from pecan.lang.pecan_ast.bool import *
 from pecan.lang.pecan_ast.prog import *
 from pecan.lang.pecan_ast.quant import *
+from pecan.lang.pecan_ast.directives import *
