+++ pecan/lang/pecan_ast/prog.py
@@ -48,6 +48,9 @@ class Call(Predicate):
         self.name = name
         self.args = args
 
+    def evaluate(self, prog):
+        return prog.call(self.name, self.args)
+
     def __repr__(self):
         return '({}({}))'.format(self.name, ', '.join(self.args))
 
@@ -57,6 +60,19 @@ class NamedPred(ASTNode):
         self.name = name
         self.args = args
         self.body = body
+
+        self.body_evaluated = None
+
+    def call(self, prog, arg_names=None):
+        if self.body_evaluated is None:
+            self.body_evaluated = self.body.evaluate(prog)
+
+        if arg_names is None:
+            return self.body_evaluated
+        else:
+            arg_vars = map(lambda name: spot.formula_ap(name), arg_names)
+            substitution = Substitution(dict(zip(self.args, arg_vars)))
+            return AutomatonTransformer(self.body_evaluated, substitution.substitute).transform()
 
     def __repr__(self):
         return '{}({}) := {}'.format(self.name, ', '.join(self.args), self.body)
@@ -84,6 +100,12 @@ class Program(ASTNode):
 
         return self
 
+    def call(self, pred_name, args=None):
+        if pred_name in self.preds:
+            return self.preds[pred_name].call(self, args)
+        else:
+            raise Exception(f'Predicate {pred_name} not found!')
+
     def __repr__(self):
         return repr(self.defs)
 
@@ -99,121 +121,4 @@ class Context:
 
     def __repr__(self):
         return '{}'.format(self.name)
-
-class Directive(ASTNode):
-    def __init__(self, name):
-        super().__init__()
-        self.name = name
-
-    def __repr__(self):
-        return '#{}'.format(self.name)
-
-class DirectiveSaveAut(ASTNode):
-    def __init__(self, filename, pred_name):
-        super().__init__()
-        self.filename = filename[1:-1]
-        self.pred_name = pred_name
-
-    def evaluate(self, prog):
-        prog.preds[self.pred_name].body.evaluate(prog).save(self.filename)
-        return None
-
-    def __repr__(self):
-        return '#save_aut({}, {})'.format(str(self.filename), self.pred_name)
-
-class DirectiveSaveAutImage(ASTNode):
-    def __init__(self, filename, pred_name):
-        super().__init__()
-        self.filename = filename[1:-1]
-        self.pred_name = pred_name
-
-    def evaluate(self, prog):
-        evaluated = prog.preds[self.pred_name].body.evaluate(prog)
-        with open(self.filename, 'w') as f:
-            f.write(evaluated.show().data) # Write the raw svg data into the file
-
-        return None
-
-    def __repr__(self):
-        return '#save_aut_img({}, {})'.format(repr(self.filename), self.pred_name)
-
-class DirectiveSavePred(ASTNode):
-    def __init__(self, filename, pred_name):
-        super().__init__()
-        self.filename = filename[1:-1]
-        self.pred_name = pred_name
-
-    def evaluate(self, prog):
-        try:
-            with open(self.filename, 'r') as f:
-                new_prog = prog.parser.parse(f.read())
-
-            with open(self.filename, 'w') as f:
-                for d in new_prog.defs:
-                    if type(d) is NamedPred and d.name == self.pred_name:
-                        self.write_pred(f, prog)
-                    else:
-                        f.write(repr(d))
-                        f.write('\n')
-        except FileNotFoundError: # No problem, just create the file
-            with open(self.filename, 'w') as f:
-                self.write_pred(f, prog)
-
-
-    def write_pred(self, f, prog):
-        for ctx in prog.context:
-            f.write(repr(DirectiveContext(ctx.name)))
-            f.write('\n')
-        f.write(repr(prog.preds[self.pred_name])) # Write the predicate definition itself into the file
-        f.write('\n')
-        for ctx in prog.context[::-1]:
-            f.write(repr(DirectiveEndContext(ctx.name)))
-            f.write('\n')
-
-    def __repr__(self):
-        return '#save_pred({}, {})'.format(repr(self.filename), self.pred_name)
-
-class DirectiveLoadPreds(ASTNode):
-    def __init__(self, filename):
-        super().__init__()
-        self.filename = filename[1:-1]
-
-    def evaluate(self, prog):
-        with open(self.filename, 'r') as f:
-            prog.parser.parse(f.read()).evaluate(prog)
-
-        return None
-
-    def __repr__(self):
-        return '#load_preds({})'.format(repr(self.filename))
-
-class DirectiveContext(ASTNode):
-    def __init__(self, context_name):
-        super().__init__()
-        self.context = Context(context_name)
-
-    def evaluate(self, prog):
-        prog.context.append(self.context)
-        return None
-
-    def __repr__(self):
-        return '#context({})'.format(self.context)
-
-class DirectiveEndContext(ASTNode):
-    def __init__(self, context_name):
-        super().__init__()
-        self.context = Context(context_name)
-
-    def evaluate(self, prog):
-        # Pop just the last context with this name
-        rev_idx = prog.context[::-1].index(self.context)
-
-        # TODO: Should we throw an error here if we don't find the context? Or just have a quick static checking phase at the beginning
-        if rev_idx >= 0:
-            prog.context.pop(len(prog.context) - rev_idx - 1) # Convert the index from an index in the reversed list to an index in the original list
-
-        return None
-
-    def __repr__(self):
-        return '#end_context({})'.format(self.context)
