+++ pecan/tools/__init__.py
@@ -1,0 +1,2 @@
+import pecan.tools.automaton_tools
+
