+++ pecan/tools/automaton_tools.py
@@ -1,0 +1,85 @@
+#!/usr/bin/env python3.6
+# -*- coding=utf-8 -*-
+
+from functools import reduce
+
+import buddy
+import spot
+
+class AutomatonTransformer:
+    def __init__(self, original_aut, formula_builder):
+        self.original_aut = original_aut
+        self.formula_builder = formula_builder
+
+    def transform(self):
+        # Build a new automata with different edges
+        bdict = spot.make_bdd_dict()
+        new_aut = spot.make_twa_graph(bdict)
+
+        aps = {}
+        for ap in self.original_aut.ap():
+            aps[ap.ap_name()] = buddy.bdd_ithvar(new_aut.register_ap(ap.ap_name()))
+
+        new_aut.set_buchi() # Set the acceptance condition to the normal Buchi acceptance condition
+        new_aut.new_states(self.original_aut.num_states())
+        new_aut.set_init_state(0)
+
+        for e in self.original_aut.edges():
+            # Convert to a formula because formulas are nicer to work with than the bdd's
+            formula = spot.formula(spot.bdd_format_formula(self.original_aut.get_dict(), e.cond))
+            cond = self.create_edge_condition(new_aut, self.formula_builder(formula))
+            new_aut.new_edge(e.src, e.dst, cond, e.acc)
+
+        return spot.minimize_obligation(new_aut)
+
+    def create_edge_condition(self, aut, new_formula):
+        return FormulaToBdd(aut, new_formula).translate()
+
+class Substitution:
+    def __init__(self, subs):
+        self.subs = subs
+
+    def substitute(self, formula):
+        if formula._is(spot.op_ap):
+            if formula.ap_name() in self.subs:
+                return self.subs[formula.ap_name()]
+            else:
+                return formula
+        else:
+            return formula.map(self.substitute)
+
+class FormulaToBdd:
+    def __init__(self, aut, formula):
+        self.aut = aut
+        self.formula = formula
+
+    def translate(self):
+        return self.run_translation(self.formula)
+
+    def build_bdd(self, kind, children):
+        if kind == spot.op_And:
+            return reduce(lambda a, b: a & b, children)
+        elif kind == spot.op_Or:
+            return reduce(lambda a, b: a | b, children)
+        elif kind == spot.op_Not:
+            return buddy.bdd_not(children[0]) # TODO: Put error check in here
+        elif kind == spot.op_tt:
+            return buddy.bddtrue
+        elif kind == spot.op_ff:
+            return buddy.bddfalse
+        else:
+            raise Exception('Unhandled formula kind: {}'.format(kind))
+
+    def run_translation(self, formula):
+        if formula._is(spot.op_ap): # op_ap is 'atomic proposition' (i.e., variable)
+            return buddy.bdd_ithvar(self.aut.register_ap(formula.ap_name()))
+        else:
+            new_children = []
+            for i in range(formula.size()):
+                new_child = self.run_translation(formula[i])
+
+                if new_child is not None:
+                    new_children.append(new_child)
+
+            return self.build_bdd(formula.kind(), new_children)
+
